requirements_ocr.txt
# OCR 정확도 향상을 위한 추가 패키지

# 기본 OCR
pytesseract>=0.3.10
opencv-python>=4.8.0

# 고급 OCR 옵션들
easyocr>=1.7.0          # 딥러닝 기반 OCR (가장 정확)
paddleocr>=2.7.0        # 표 인식 우수

# 표 추출 전용
camelot-py[cv]>=0.11.0  # PDF 표 추출
tabula-py>=2.8.0        # 빠른 표 추출
pdfplumber>=0.10.0      # 레이아웃 유지

# 이미지 처리
Pillow>=10.0.0
numpy>=1.24.0
